pyproject.toml
[tool.poetry]
name = "agent-governance-hub"
version = "0.1.0"
description = "A lightweight, open-source governance middleware for LLM agents"
readme = "README.md"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "agents"}, {include = "api"}, {include = "config"}, {include = "governance"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
pyyaml = "^6.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
langchain = "^0.1.0"
langchain-community = "^0.0.20"
langchain-openai = "^0.0.5"
sentence-transformers = "^2.2.0"
qdrant-client = "^1.7.0"
openai = "^1.0.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
httpx = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
